Update the OpenDS setup process to add a "Please wait while the server initializes...." message at the beginning of the process which can make any delays introduced by the initial configuration and schema processing more palatable to the end user.
This also updates the setup.sh shell script so that it will attempt to use the
client VM under certain circumstances, since we have found that the client VM
does significantly reduce the startup time in many cases.  It should continue
to work properly in other VMs that may not support the "-client" option.

OpenDS Issue Number:  594


git-svn-id: https://svn.forgerock.org/opendj/trunk@197 41b1ffd8-f28e-4786-ab96-9950f0a78031