Provide a mechanism for managing file permissions.  On UNIX-based systems where the use of exec is allowed, it will use the underlying chmod utility to set file permissions.  On other systems, if Java 6 is available then the new methods in the java.io.File class will be used.  If neither option is available, then it will not be possible to manage file permissions.
This update also includes a mechanism that can be used to allow administrators
to preven the server from using exec to run commands on the underlying system.
Note that this is more a configuration option than an actual enforcement,
although it could be possible to configure the security manager to explicitly
forbid this.

OpenDS Issue Number:  601


git-svn-id: https://svn.forgerock.org/opendj/trunk@227 41b1ffd8-f28e-4786-ab96-9950f0a78031