Update the Directory Server build process so that the test target will create a minimized installation that will allow the server to be started for those tests that need access to the configuration, schema, or any other aspect of the running server.  Also, update the test cases to make use of this capability when necessary.
This is intended to replace the current *Fixture elements, which don't create a
sufficient directory structure to be able to start the server.  Further, this
uses a directory structure that will not be deleted when the tests are complete
in the event that further investigation into a failure is necessary.


git-svn-id: https://svn.forgerock.org/opendj/trunk@266 41b1ffd8-f28e-4786-ab96-9950f0a78031