Fix for  [Issue 635]  NullPointerException when trying to access non existing entry.  Reviewed by gbellato.
The synchronization post-operation plugin did not handle the case where an operation had not been assigned a change number.  A unit test is added for this.  There was also an error in the increment of numReplayedPostOpCalled in the case of a non-synchronization or unsuccessful operation, since that field is documented as "the number of updates replayed successfully by the synchronization".

git-svn-id: https://svn.forgerock.org/opendj/trunk@597 41b1ffd8-f28e-4786-ab96-9950f0a78031