Fix for  [Issue 846] DN embedded in a monitor instance name.  Reviewed by gbellato.
The fix is to avoid constructing an RDN using string concatenation of the attribute type and value, when the value could contain characters that need to be quoted inside of a DN.  I looked through the rest of the code at uses of DN.decode() and RDN.decode() and did not find any other places where the same mistake has been made.

The changes include updated unit tests to catch this issue.

git-svn-id: https://svn.forgerock.org/opendj/trunk@600 41b1ffd8-f28e-4786-ab96-9950f0a78031