issue 508 These changes implement a window mechanism in the sycnhronization protocol.
Up to now the flow control mechanism used by the synchronization
was  the TCP flow control mechanism. However, since TCP is not aware about
the type of the synchronization mechanism this was not allowing sending
of ACK messages when the TCP connection was saturated.
This was also preventing the implementation of the prioritized synchronization.

With these changes the TCP windows are set to a very large value and the
flow control is based on a configurable window size on the changelog servers
and on the LDAP servers.

These changes also add monitoring informations for the current and max window sizes.

I also took the opportunity to remove most the static variables and methods that were
preventing multi instantiation of the Changelog class.

I have also added tests for the incoding/decoding of ServerStartMessage and
ChangelogStartMessage, WindowMessage and for testing the window mechanism.

Also add the possibility to choose the port number used by the LDAP server when running
the unit test using property : org.opends.server.LdapPort
This can be usefull for debugging purpose.

reviewed by Daniel


git-svn-id: https://svn.forgerock.org/opendj/trunk@700 41b1ffd8-f28e-4786-ab96-9950f0a78031