Add two new server properties:
- org.opends.server.LockDirectory -- Specifies the path to the directory that
  should hold the server lock files

- org.opends.server.SchemaDirectory -- Specifies the path to the directory that
  should hold the server schema files

These properties can help make it easier for applications to embed OpenDS and
use a filesystem layout that is different from the default layout used by the
server.  If either of these properties is not provided, then the default
location will be used (locks for the lock directory, and config/schema for the
schema directory).

OpenDS Issue Number:  649


git-svn-id: https://svn.forgerock.org/opendj/trunk@914 41b1ffd8-f28e-4786-ab96-9950f0a78031