put back the 980 revision that had been removed because of a regression.  - fix the regression (schema problem)  - disabled the ChangelogTest.stopChangelog() test that appear to cause    some hangs.
----------

This change includes a set of new unit tests dedicated to the startup phase
of the synchronization server (former changelog server) and the connection
phase of client LDAP server to this synchronization server.
These tests cover various cases of this connection checking that the adequate
changes are then sent to the client.

This change also include the fix for :
issue 1155 : synchronization server miss one change when a new LDAP server is added
to the topology

git-svn-id: https://svn.forgerock.org/opendj/trunk@996 41b1ffd8-f28e-4786-ab96-9950f0a78031