fix for issue 1212 namingConflict test trigger an assert from the static group
When it is enabled, the synchronization code can sometimes decide to abandon
an add operation. In such case the handleConflictResolution() method
of the SynchronizationProvider will return a code indicating to stop the
OperationProcessing.
However when doing this, the ChangeListener are still invoked
at the end of the AddOperation processing.
Since the operation has not been processed, the ChangeListeners are given
a null entry parameter but the StaticGroup code that runs as a ChangeListener
makes the assumptions that the entry is always non null at this point.
This currently causes the following assert when running the test
synchronization.UpdateOperationTest.namingConflicts()

The change notification listeners will only be invoked if the result
code for the operation is "SUCCESS", and at that point in the processing
it should only be SUCCESS if the update was actually applied.  If the
synchronization code decided to abort the operation the
the provider.handleConflictResolution() therefore now set
a Cancel Result code.







git-svn-id: https://svn.forgerock.org/opendj/trunk@1201 41b1ffd8-f28e-4786-ab96-9950f0a78031