Fix for issue 1285.
The problem was caused because in the class InstallDS the check for the port was not performed when the user already provided the port through the command line (see first diffs in InstallDS).

The rest of the diffs consist of factorization of code to be shared between the quicksetup classes and the class org.opends.server.tools.InstallDS.  Some methods from the Utils class of quicksetup have been moved to the SetupUtils class; these methods allow to know if a given port is priviledged or not, if we can use it to configure the server and to know in which Operating System the JVM is running.   The operating system methods use the enumeration OperatingSystem (which has been slightly modified to not have dependencies on other classes and to be included in quicksetup.jar).

Extra: use the class LockFileManager to know if the server is running or not.  The use of this class does not impact the Web Start Installer as we only want to know if the server is running or not once we have tried to start it (and so OpenDS.jar has already been loaded by the class loader).


git-svn-id: https://svn.forgerock.org/opendj/trunk@1272 41b1ffd8-f28e-4786-ab96-9950f0a78031