issue 504,505,507 : res-synchronization after restore or import
The synchronization protocol already have some re-synchronization capabilities
that allows the synchronization server to detect where to start from the changelog
when a new LDAP server opens a session.

These changes implements the ability to do this when a restore or
an import is done when the LDAP server is online and already connected to a
synchronization server.

To achieve this the synchronization code, on notification from the core server
that a restore or an import is starting, close the session to the synchronization
server and restart it when the restore or import is finished.

The synchronization protocol then does its job of negociating where to start
from the changelog and resynchronizing the LDAP Server.

The core server does not yet provide the notification to the changelog server
but here if the synchronization part of the code along with some test for it.


git-svn-id: https://svn.forgerock.org/opendj/trunk@1296 41b1ffd8-f28e-4786-ab96-9950f0a78031